requirements.txt
nltk==3.2.5
GitPython==2.1.8
argparse==1.4.0
nltk==3.2.5
